requirements-dicom-study-demo.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
pydicom>=2.4.0
Pillow>=10.0.0
numpy>=1.24.0
pylibjpeg>=2.0.0
pylibjpeg-libjpeg>=2.1.0
pylibjpeg-openjpeg>=2.2.0
